# Optimize journal queries