--- a/opengever/journal/browser.py
+++ b/opengever/journal/browser.py
@@ -1,11 +1,13 @@
 
 from Acquisition import aq_inner
+from ftw.journal.config import JOURNAL_ENTRIES_ANNOTATIONS_KEY
+from ftw.journal.interfaces import IAnnotationsJournalizable
+from ftw.journal.interfaces import IWorkflowHistoryJournalizable
 from Products.Five.browser import BrowserView
-from zope.annotation.interfaces import IAnnotations, IAnnotatable
+from zope.annotation.interfaces import IAnnotatable
+from zope.annotation.interfaces import IAnnotations
 from zope.interface import alsoProvides
 
-from ftw.journal.interfaces import IAnnotationsJournalizable, IWorkflowHistoryJournalizable
-from ftw.journal.config import JOURNAL_ENTRIES_ANNOTATIONS_KEY
 
 class JournalHistory(BrowserView):
     """ BrowserView listing the journal history