--- a/opengever/journal/browser.py
+++ b/opengever/journal/browser.py
@@ -10,8 +10,7 @@ from zope.interface import alsoProvides
 
 
 class JournalHistory(BrowserView):
-    """ BrowserView listing the journal history
-    """
+    """BrowserView listing the journal history."""
 
     def __init__(self, context, *args, **kwargs):
         self.context = aq_inner(context)
@@ -19,9 +18,6 @@ class JournalHistory(BrowserView):
         super(JournalHistory, self).__init__(context, *args, **kwargs)
 
     def data(self):
-        """
-        """
-
         if IAnnotationsJournalizable.providedBy(self.context):
             annotations = IAnnotations(self.context)
             return annotations.get(JOURNAL_ENTRIES_ANNOTATIONS_KEY, [])