@Rotonen 2017-05-22
@deiferni ready for rereview.